Add CRUD methods and GetFullName to Person class
Added a method `GetFullName` that returns the full name of the person by concatenating the `FirstName` and `LastName` properties.
Added a method `CreatePerson` that initializes the `FirstName`, `LastName`, and `DateOfBirth` properties of the `Person` class.
Added a method `UpdatePerson` that updates the `FirstName`, `LastName`, and `DateOfBirth` properties of the `Person` class.
Added a method `DeletePerson` that sets the `FirstName` and `LastName` properties to `null`.

#2